% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazip.R
\name{datazip}
\alias{datazip}
\title{Compress and Save Data frame Code}
\usage{
datazip(df, output = NULL)
}
\arguments{
\item{df}{A dataframe to be compressed and optionally saved.}

\item{output}{A character string specifying the file path to save the output.
If \code{NULL}, the compressed representation is printed to the console.
Supported formats are \code{.rds} and \code{.r}.}
}
\value{
Returns the modified dataframe invisibly.
}
\description{
This function compresses a data.frame by removing spaces from column names and
formatting its \code{dput()} output for compact representation. The function can
also save the data.frame as an RDS or R script file.
}
\examples{
if(!require(remotes)) install.packages("remotes")
if (!requireNamespace("datazip", quietly = TRUE)) {
remotes::install_github("agronomy4future/datazip", force= TRUE)
}
library(remotes)
library(datazip)

df= data.frame(x= 1:5, y= letters[1:5])
# convert data to code
datazip(df) # Print compressed representation
datazip(df, output="data1.rds") # Save as RDS
datazip(df, output="data2.r") # Save as R script

# import code to data
df_loaded_rds= dataunzip("df.rds")  # Load from RDS file
df_loaded_r= dataunzip("df.r")      # Load from R script file
print(df_loaded_rds)
print(df_loaded_r)
}
